+++ src/main/java/insane96mcp/progressivebosses/module/elderguardian/feature/AttackFeature.java
@@ -1,0 +1,76 @@
+package insane96mcp.progressivebosses.module.elderguardian.feature;
+
+import insane96mcp.insanelib.base.Feature;
+import insane96mcp.insanelib.base.Label;
+import insane96mcp.insanelib.base.Module;
+import insane96mcp.progressivebosses.setup.Config;
+import net.minecraft.entity.monster.ElderGuardianEntity;
+import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.event.entity.living.LivingHurtEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+@Label(name = "Attack", description = "More damage and attack speed based off Elder Guardians Defeated")
+public class AttackFeature extends Feature {
+
+	private final ForgeConfigSpec.ConfigValue<Double> bonusDamageConfig;
+	private final ForgeConfigSpec.ConfigValue<Integer> attackDurationReductionConfig;
+
+	public double bonusDamagePer = 0.2d;
+	public int attackDurationReduction = 25;
+
+	public AttackFeature(Module module) {
+		super(Config.builder, module);
+		this.pushConfig(Config.builder);
+		bonusDamageConfig = Config.builder
+				.comment("Percentage Bonus damage per defeated Elder Guardian.")
+				.defineInRange("Bonus Damage per Elder Guardian Defeated", bonusDamagePer, 0d, 1d);
+		attackDurationReductionConfig = Config.builder
+				.comment("How many ticks faster will Elder Guardian attack (multiplied by defeated Elder Guardians). Vanilla Attack Duration is 60 ticks (3 secs)")
+				.defineInRange("Attack Duration Reduction per Elder Guardian Defeated", attackDurationReduction, 0, 60);
+		Config.builder.pop();
+	}
+
+	@Override
+	public void loadConfig() {
+		super.loadConfig();
+		this.bonusDamagePer = this.bonusDamageConfig.get();
+		this.attackDurationReduction = this.attackDurationReductionConfig.get();
+	}
+
+	@SubscribeEvent
+	public void onDamageDealt(LivingHurtEvent event) {
+		if (event.getEntity().getEntityWorld().isRemote)
+			return;
+
+		if (!this.isEnabled())
+			return;
+
+		if (this.bonusDamagePer == 0d)
+			return;
+
+		if (!(event.getSource().getTrueSource() instanceof ElderGuardianEntity))
+			return;
+
+		ElderGuardianEntity elderGuardian = (ElderGuardianEntity) event.getSource().getTrueSource();
+
+		int elderGuardiansNearby = elderGuardian.world.getEntitiesInAABBexcluding(elderGuardian, elderGuardian.getBoundingBox().grow(48d), entity -> entity instanceof ElderGuardianEntity).size();
+		if (elderGuardiansNearby == 2)
+			return;
+
+		float bonusDamage = (float) (this.bonusDamagePer * (2 - elderGuardiansNearby));
+
+		event.setAmount(event.getAmount() * (1f + bonusDamage));
+	}
+
+	private static final int BASE_ATTACK_DURATION = 60;
+
+	public int getAttackDuration(ElderGuardianEntity elderGuardian) {
+		if (!this.isEnabled() || this.attackDurationReduction == 0)
+			return BASE_ATTACK_DURATION;
+		int elderGuardiansNearby = elderGuardian.world.getEntitiesInAABBexcluding(elderGuardian, elderGuardian.getBoundingBox().grow(48d), entity -> entity instanceof ElderGuardianEntity).size();
+		if (elderGuardiansNearby == 2)
+			return BASE_ATTACK_DURATION;
+
+		return BASE_ATTACK_DURATION - ((2 - elderGuardiansNearby) * this.attackDurationReduction);
+	}
+}